contrib/jmoiron/sqlx: update to new form and add new methods
