--- contrib/jmoiron/sqlx/pq_test.go
@@ -1,41 +1,0 @@
-package sqlx
-
-import (
-	"log"
-	"testing"
-
-	"github.com/DataDog/dd-trace-go/contrib/database/sql/sqltest"
-	"github.com/DataDog/dd-trace-go/tracer"
-	"github.com/DataDog/dd-trace-go/tracer/tracertest"
-	"github.com/lib/pq"
-)
-
-func TestPostgres(t *testing.T) {
-	trc, transport := tracertest.GetTestTracer()
-	dbx, err := OpenTraced(&pq.Driver{}, "postgres://postgres:postgres@127.0.0.1:5432/postgres?sslmode=disable", "postgres-test", trc)
-	if err != nil {
-		log.Fatal(err)
-	}
-	defer dbx.Close()
-
-	testDB := &sqltest.DB{
-		DB:         dbx.DB,
-		Tracer:     trc,
-		Transport:  transport,
-		DriverName: "postgres",
-	}
-
-	expectedSpan := &tracer.Span{
-		Name:    "postgres.query",
-		Service: "postgres-test",
-		Type:    "sql",
-	}
-	expectedSpan.Meta = map[string]string{
-		"db.user":  "postgres",
-		"out.host": "127.0.0.1",
-		"out.port": "5432",
-		"db.name":  "postgres",
-	}
-
-	sqltest.AllSQLTests(t, testDB, expectedSpan)
-}
